add emmet snippets, expandable with shift enter
this commit also removes a hacky luasnip jump resetter that helped with wonky jumps for latex snippets in markdown. it sadly conflicted with the emmet snippets sometimes, but changing luasnip functionality so that it only jumps forward if locally jumpable seems to mostly fix the issue, and it is less hacky and seems more robust and idiomatic.