Pin Iosevka with a pinned version of nixpkgs
Even if we pin the package with an override, it will still want to
rebuild itself upon a flake update for some reason. This defeats the
whole purpose; I just wanted to pin the package to prevent it from
updating all the time, not because I literally want a specific version.
It seems that pinning it this way eliminates the redundant builds and
the usually upgrading. Nice!